+++ lib/res/utils/web_browser_info_stub.dart
@@ -1,0 +1,3 @@
+bool get isSafariWeb => false;
+
+bool get isIOSSafariWeb => false;